+++ src/main/java/net/yxiao233/ifeu/api/block/renderer/IFEUStructureEntityRendererJS.java
@@ -1,5 +1,6 @@
 package net.yxiao233.ifeu.api.block.renderer;
 
+import com.hrznstudio.titanium.block.RotatableBlock;
 import com.mojang.blaze3d.vertex.PoseStack;
 import net.minecraft.client.renderer.MultiBufferSource;
 import net.minecraft.client.renderer.blockentity.BlockEntityRenderer;
@@ -8,32 +9,39 @@ import net.minecraft.core.BlockPos;
 import net.minecraft.core.Direction;
 import net.minecraft.world.level.Level;
 import net.minecraft.world.level.block.Block;
+import net.minecraft.world.level.block.entity.BlockEntity;
 import net.minecraft.world.level.block.state.BlockState;
-import net.yxiao233.ifeu.api.block.entity.IFEUStructureProcessingTile;
+import net.minecraft.world.level.block.state.properties.BlockStateProperties;
 import net.yxiao233.ifeu.api.structure.MultiBlockStructure;
 import net.yxiao233.ifeu.common.utils.RendererProvider;
 
 import java.util.List;
 
-public class IFEUStructureEntityRenderer implements BlockEntityRenderer<IFEUStructureProcessingTile<?>> {
-    private volatile MultiBlockStructure structure;
+public class IFEUStructureEntityRendererJS implements BlockEntityRenderer<BlockEntity> {
+    private final MultiBlockStructure structure;
     private final int tick = 200;
-    public IFEUStructureEntityRenderer(BlockEntityRendererProvider.Context context) {
+    private int time = 0;
+    public IFEUStructureEntityRendererJS(BlockEntityRendererProvider.Context context, MultiBlockStructure structure) {
+        this.structure = structure;
     }
     @Override
-    public void render(IFEUStructureProcessingTile<?> entity, float v, PoseStack poseStack, MultiBufferSource multiBufferSource, int i, int i1) {
-        if(entity.isShouldRenderer() && entity.direction != null){
+    public void render(BlockEntity entity, float v, PoseStack poseStack, MultiBufferSource multiBufferSource, int i, int i1) {
+        BlockState blockState = entity.getBlockState();
+        boolean hasDirection = blockState.hasProperty(BlockStateProperties.HORIZONTAL_FACING);
+        boolean hasSpecialDirection = blockState.hasProperty(RotatableBlock.FACING_HORIZONTAL);
+        Direction direction = null;
+        if(hasDirection){
+            direction = blockState.getValue(BlockStateProperties.HORIZONTAL_FACING);
+        }else if(hasSpecialDirection){
+            direction = blockState.getValue(RotatableBlock.FACING_HORIZONTAL);
+        }
+        if(direction != null){
             Level level = entity.getLevel();
             if(level == null){
                 return;
             }
-
-            structure = getStructure(entity);
-
-            Direction direction = entity.direction;
             BlockPos machinePos = entity.getBlockPos();
-            boolean renderFull = entity.renderFull;
-            entity.increaseTime(1);
+            time ++;
 
             var poses = structure.getRenderStructure(direction,machinePos);
             RendererProvider provider = new RendererProvider(poseStack,multiBufferSource,machinePos);
@@ -42,13 +50,13 @@ public class IFEUStructureEntityRenderer implements BlockEntityRenderer<IFEUStru
             poses.getLeft().forEach(pair ->{
                 BlockPos pos = pair.getLeft();
                 BlockState state = pair.getRight().defaultBlockState();
-                render(pos,state,provider,minY,!level.getBlockState(pos).is(pair.getRight()),renderFull);
+                render(pos,state,provider,minY,!level.getBlockState(pos).is(pair.getRight()));
             });
 
             poses.getRight().forEach(pair ->{
                 BlockPos pos = pair.getLeft();
                 List<Block> blocks = pair.getRight().getLeft();
-                int index = (int) Math.floor((double) entity.getTime() / tick);
+                int index = (int) Math.floor((double) time / tick);
                 while(index >= blocks.size()){
                     index = index - blocks.size();
                 }
@@ -56,30 +64,19 @@ public class IFEUStructureEntityRenderer implements BlockEntityRenderer<IFEUStru
                     index = 0;
                 }
                 BlockState state = blocks.get(index).defaultBlockState();
-                render(pos,state,provider,minY,!level.getBlockState(pos).is(pair.getRight().getRight()),renderFull);
+                render(pos,state,provider,minY,!level.getBlockState(pos).is(pair.getRight().getRight()));
             });
         }
     }
 
-    private void render(BlockPos pos, BlockState state, RendererProvider provider, int minY, boolean shouldRender, boolean renderFull){
+    private void render(BlockPos pos, BlockState state, RendererProvider provider, int minY, boolean shouldRender){
         if(shouldRender){
-            if(renderFull){
-                provider.renderSingleBatchedGhostBlock(pos,state);
-            }else if(pos.getY() != minY){
+            if(pos.getY() != minY){
                 //为了不让idea提示,没有任何实际用处
                 int var0 = 0;
             }else{
                 provider.renderSingleBatchedGhostBlock(pos,state);
-
             }
-        }
-    }
-
-    public MultiBlockStructure getStructure(IFEUStructureProcessingTile<?> entity){
-        if(structure == null){
-            return entity.multiBlockStructure();
-        }else{
-            return structure;
         }
     }
 }