+++ src/main/java/net/yxiao233/ifeu/common/compact/kubejs/events/IFEUEvents.java
@@ -6,5 +6,6 @@ import dev.latvian.mods.kubejs.event.EventHandler;
 public class IFEUEvents {
     public static final EventGroup GROUP = EventGroup.of("IFEUEvents");
 
-    public static final EventHandler STRUCTURES = GROUP.startup("structureModify", () -> IFEUStructuresEvent.class);
+    public static final EventHandler STRUCTURES = GROUP.startup("structureModify", () -> IFEUStructureModifyJS.class);
+    public static final EventHandler CREATE = GROUP.startup("structureRender", () -> IFEUStructureRenderJS.class);
 }